gRPC tests: run server once, use it in all tests
